move dirs, add claim.rs
