+++ program/src/check_pda.rs
@@ -1,0 +1,22 @@
+use crate::error::ClaimError;
+use solana_program::{msg, program_error::ProgramError, pubkey::Pubkey};
+
+pub fn check_claim_pda(
+    seeds: &[&[u8]],
+    claim_program: &Pubkey,
+    airdrop_account: &Pubkey,
+) -> Result<(), ProgramError> {
+    let derived_pda =
+        Pubkey::create_program_address(seeds, claim_program).expect("Invalid PDA seeds.");
+
+    if derived_pda != *airdrop_account {
+        msg!(
+            "Invalid airdrop PDA provided. Expected: {}. Found: {}.",
+            derived_pda,
+            airdrop_account
+        );
+        return Err(ClaimError::InvalidPDA.into());
+    }
+
+    Ok(())
+}
