+++ program/src/constants.rs
@@ -1,0 +1,3 @@
+use solana_program::{pubkey, pubkey::Pubkey};
+
+pub const CTOKEN_PROGRAM_ID: Pubkey = pubkey!("cTokenmWW8bLPjZEBAUgYy3zKxQZW6VKi7bqNFEVv3m");
